Added missing doc and copyright notices. Started a bug refactor on how to handle manufacturability. Still halfway there, but enough for everyone to see where I'm going + compile.
The idea is to get the manufacturability data separate from the object itself (therefore preventing lots of useless values being stored in memory and granting more direct access to manufacturable items through s eparate structure). Also added a better method to detect the type of manufacture it is, and some missing methods to allow to load it.

git-svn-id: https://aoxp.googlecode.com/svn/trunk@281 4a1aa3fa-7d1d-11de-9afa-c7a50e5ca1b3